OverlayConfig: avoid implementing OverlayConfigDialogNew on OS X Universal.
The OverlayConfigDialogNew function is only used by the ConfigRegistrar
just below its declaration.

That ConfigRegistrar is only created on system that aren't OS X Universal.
Now that we build with -Werror, the fact that OverlayConfigDialogNew is
unused on OS X Universal is causing trouble.

To fix the warning about OverlayConfigDialogNew being unused on OS X
Universal, we simply move it into the if !defined(USE_MAC_UNIVERSAL)
block, since it is not used anywhere else.