Log: trigger scroll-to-bottom for a QAbstractTextDocumentLayout's update() and documentSizeChanged() signals.
For me, this fixes the bad scroll-to-bottom behavior we have had
in some earlier 1.2.x versions.

My test case is:

 - Connect two clients to a test server
 - Send various multi-line text messages using the chat bar.
 - Send a single image (data:// URL)

This would consistently fail to scroll to the bottom of the log view
in 1.2.10.

With this patch applied, it works as it should.

Not a pretty solution, but it works.