LogTextBrowser: move all scroll-to-bottom logic to the LogTextBrowser.
Instead of trying to handle it in the LogDocument, this commit moves
the scroll-to-bottom logic into the LogTextBrowser class.

We now track when the user moves the scrollbar.	If the user moves the
scrollbar to the very bottom, or if the user has not touched the scrollbar
at all, we will now scroll to the bottom of the log view when the content
of the log view is changed.

If the user has moved the scroll position -- for example if the user
is reading an old message -- we will keep the user in that position,
even if new entries are added to the log view.

Fixes mumble-voip/mumble#1969