GlobalShortuct_win: warn if a device takes more than 1 second to poll via DirectInput.
We've had reports of various DACs misbehaving, causing Mumble to freeze
for a long time when they use global shortcuts.

See for example mumble-voip/mumble#1880 and mumble-voip/mumble#1977.

This is not a fix, but it gives users who face this problem a hint if they
look in their Mumble log.