GlobalShortcut_win, XboxInput: implement native XInput support in GlobalShortcut_win.
This commit adds native support for XInput to GlobalShortcut on Windows.

Most things XInput-related is named XboxInput in an attempt to reduce
confusion with X11's XInput.

This commit adds a new class, XboxInputLibrary, which is a thin wrapper
class for handling the XInput DLL.

XboxInputLibrary supports various versions of the XInput DLL, as well as
the non-public XInputGetStateEx function (ordinal 100) -- if it is
available in the loaded XInput DLL. XInputGetStateEx makes it possible to
query the state of the guide button, allowing it to be used as a global
shortcut in Mumble.

The most suitable GetState function is chosen by XboxInputLibrary and is
exposed simply via the GetState member variable.

This commit does nothing to prevent DirectInput from *also* getting input
from any connected Xbox controllers. However, no ill effects have been
observed yet.

For now, the result simply is that -- for the buttons that DirectInput
supports -- we get a button click from both engines. For example,
pressing A yields the equivalent of "DirectInput:A + Xbox1:A" as the
shortcut name, because both engines handled the button press.

In the future, we should blacklist the DirectInput device if XboxInput
is enabled to rule out any future incompatibilities.

The current implementation treats a shortcut from the first Xbox
controller as a different shortcut than the same button from a
potential second, third or fourth Xbox controller connected to
the system. As a result, if a button is bound on the first controller
in one session, Mumble will only respond if that button is pressed on
the first controller in future sessions.

Fixes mumble-voip/mumble#1995