ConnectDialog: disable ping and host lookups when using a proxy.
This commit adds four flags to the ConnectDialog. One that allows
pings, one that allows system hostname lookups, one that allows
Bonjour host lookups, and one that allows the use of filters in
the ConnectDialog.

By default, they are all enabled.

However, if you're using a proxy, all will be disabled.

The commit also updates the ConnectDialog UI to hide
the user count and ping columns when pings are not enabled.
This is because it is generally not possible to ping via the
usual UDP mechanism when using Mumble through a proxy.

Also, since it is not possible to ping servers, the filters
that are usually available for the ConnectDialog are also
disabled. When using a proxy, all servers are shown.

As a consequence of this commit, Mumble will now allow you to
connect to a server for which a hostname has not yet been
resolved if you are using a proxy.

Previously, for hosts that could not be resolved via the regular
resolvers in Mumble (system resolver, Bonjour, ...), Mumble
would not be able to connect. This is because Mumble would disable
the Connect button if no hostname(s) could be resolved -- or were
not yet resolved.

With this commit, Mumble will now allow connections to those
hosts, and let the proxy will do the hostname resolving. This
allows, among other things, Mumble to connect to Tor .onion
hosts when using Tor as a proxy.

Fixes mumble-voip/mumble#1033
Fixes mumble-voip/mumble#1812