Murmur: allow both bitmasked version and version string when setting 'suggestversion' via RPC.
This commit changes the logic for setting the 'suggestversion' config
option via RPC to automatically figure out whether the passed-in value
is a bitmasked version or a version string.

If it is a version string, it will convert the version string to the
bitmasked format, which is what is used internally.

Prior to this, 'suggestversion' via RPC only allowed a bitmasked
version. This was confusing users, since the .ini file requires a
a version string.

Fixes mumble-voip/mumble#989