About: change 'The Mumble team' to 'The Mumble Developers', as used in our license.
Our LICENSE file references:

    "[...] Neither the name of the Mumble Developers nor
     the names of its contributors [...]"

As discussed in mumble-voip/mumble#1513, we should use that when
referencing the authors of Mumble.

This commit also makes the string untranslatable. This is
intentional, as "The Mumble Developers" identifies the name
we use for the project in a copyright context.

In fact, the whole Copyright string in the About dialog is now
marked as untranslatable. I think this is sensible, as our license
text itself is in English, and the meaning of the word "Copyright"
should be widely recognized even by non-English speakers. Referencing
local equivalents of Copyright law could cause confusion regarding
the legal status of Mumble.

Updates mumble-voip/mumble#1513