Mumble: implement lock file mechanism for Windows.
The RPC/DBus-based method for ensuring that only a single instance of
Mumble is running is racy.

If one of the processes has started it's RPC listener, it works fine.
But if both processes are started at almost the same time, it's very
likely that neither process has started listening for RPC. And the
exclusivity check fails.

In order to be completely sure that only a single Mumble instance is
running, we now take a lock file.

We open a lock file file in the data directory (for Windows, that's
%APPDATA%\Mumble). The file has sharing disabled, so as long as the
process that created the file is running, no other process can access
it.

If a second instance of Mumble is launched, and the file is inaccessible,
that instance terminates itself.

It is, of course, still possible to override the single instance
limitation by passing the '-m' flag to Mumble.

Fixes mumble-voip/mumble#1797