compiler.pri: fix architecture checks for force-x86-toolchain and force-x86_64-toolchain builds.
Fixes the build when using the win32-static buildenv (32-bit), due to a
linker warning (turned error due to -Wx) because the XP subsytem (5.01)
is not valid on the x64 compiler.

Technically, this change implements a new qmake variable in
compiler.pri, MUMBLE_ARCH. By default, it's set to the same
value as QMAKE_TARGET.arch, but when using force-x86-toolchain
or force-x86_64-toolchain, it is set to the cross compiler's
architecture.