Do not enable warnings-as-errors by default.
While we would like to have warnings always be treated
as errors, enabling it will surely cause headaches for
packagers. Different compiler versions behave differently,
and it could potentially turn out top be very annoying for
distro packagers if they try to build an older version of
Mumble using a compiler that we had not anticipated being
used for building Mumble (a newer version of GCC, for
example).

Because of that, we disable warnings-as-errors by default.

Our own binary builds will, of course, enable it
warnings-as-errors to ensure we don't miss anything.

The CI system will also always use it. That should
hopefully also ensure we haven't overlooked anything
on platforms where we don't provide binary builds
ourselves.

Fixes mumble-voip/mumble#2113