plugins/ql: Fix host address string and add comments
There are two memory addresses: one with the obfuscated IP address and one with the true one. Initially I used the first one; when you join through the console it has the true IP address, but when you join through the server browser it shows the obfuscated one: this is a problem. I decided to use the second memory address, that always shows the true IP address.
Reason of the string change: I noticed that the second memory address uses "loopback" instead of "localhost" used by the first, when on a local server.