scripts/generate-AUTHORS.py: use pyuca for sorting to get deterministic output on all platforms.
My own rule of thumb for scripts in our repo is to keep the
number dependencies down. In practice, this means I always
strive to only use the standard library.

In this case, it's not that easy.

The existing code sorted sufficiently on Windows.

However, when run on Unix-like systems, it produces odd, and
to my mind, unexpected sorting behavior. (Such as ignoring spaces,
and sorting 'Hey You' after 'Heyh You'.)

I suppose the sort order is a matter of preference.

But the non-determinism of the script's output isn't.
If we don't fix this, we'll get noisy diffs once in a while,
which isn't very nice.

This commit changes the script to use 'pyuca' to
do the sorting. This is a pure Python module, so
it's easy to install via pip on all OSes.