mumble_plugin_win32.h: add explicit procptr32_t and procptr64_t pointer types and add peekProc overloads for them.
Right now, a lot of our plugins expect to be able to use
    peekProc<BYTE *>(someAddr);
to read a pointer value from a game.

When the plugin is built as 32-bit, and the game is 32-bit, all is fine.
The peekProc operation above will read a 32-bit pointer from the game's
memory.

However, when the plugin is built as 64-bit, and the game is 32-bit,
things start to break down, because the `BYTE *` type suddenly isn't
32-bit wide anymore.

This commit tries to remedy that by introducing explicit types for
32-bit and 64-bit pointers.

We expect that 32-bit games have one plugin, and 64-bit games have
another plugin (see for example bf4 and bf4_x86).

With this change, plugins targetting 32-bit games should use the
new procptr32_t type when querying pointers from the game. Plugins
targetting 64-bit games should use procptr64_t.

This commit also adds overloads for peekProc() to be able to use
procptr32_t and procptr64_t as base addresses.

This commit also adds pModule32 and pModule64 variables to
mumble_plugin_win32.h. These are the base address of the game's
executable in the game's address space. Plugin authors should use
the the pModule32 variable for 32-bit games, and the pModule64 variable
when targetting 64-bit games.

Once we've transitioned the whole repository over to use these new
explicit pointer types, we can drop the peekProc() overloads that
take VOID *. We can also drop the 'VOID *'-typed pModule.