Overlay: add launcher filter.
The launcher filter is a step before the usual whitelist/blacklist.

Users can maintain a list of allowed "launcher" processes, and Mumble
will only inject the overlay into the process if that process's parent
is part of this list of allowed launchers.

The idea is that most games these days run through a launcher.

Either the various game stores, or per-title launchers for some
games.

The hope is that this will remove a lot of the annoyances with
the overlay feature -- at least for non-technical users. We've
often had bug reports of Mumble injecting itself into things like
browsers, producitiy apps, and so on. This change should ensure
that, by default, the overlay only injects itself into games.