build: add pkgconfig.pri and must_pkgconfig qmake function.
This commit adds a new helper function, must_pkgconfig.

It is intended as a replacement for:

    PKGCONFIG *= openssl

The must_pkgconfig function performs error checking,
and fails if the package is not available.

This is to avoid situations mumble-voip/mumble#2309
where the build environment is not set up correctly
in some way.

When using must_pkgconfig, we fail early with a
clear error message, instead of failing with
a cryptic build-time error later on.