TextMessage, RichTextEditor: accept TextMessage dialog when Ctrl-Enter is pressed.
A new signal, accept(), is added on RichTextEditor.
This signal is emitted when Ctrl-Enter is pressed in a RichTextEditor.

This commit rips out an old non-working event filter TextMessage
that was supposed to accept the dialog when the user pressed
Enter. However, this code didn't work anymore because the
QTextEditor widgets of RichTextEditor swallowed all the KeyPress
events.

Finally, in the TextMessage class, the accept() signal from RichTextEditor
is connected to the accept() of TextMessage's QDialog, allowing
users to accept the dialog with Ctrl-Enter.

Fixes mumble-voip/mumble#2294