GlobalShortcut_win: when adding a DirectInput device, log its dwDevType.
We've previously resorted to blacklisting devices that cause hangs and
other issues.

The goal of logging dwDevType here is to be able to diagnose such issues
in the future. It would be interesting to know what such devices advertise
themselves as to DirectInput.

Perhaps, once we have more data, we can limit the amount of devices we
query, instead of querying all devices all the time.