Log: check for QSizeF::isValid() in Log::validHtml().
It is possible for QTextDocument::size() to return
invalid sizes. For example, very big img tags with
sizes such as 33554451x33554451 cause QTextDocument
to return a QSizeF with a negative width or height.

This commits adds a call to QSizeF::isValid() to ensure
we treat QSizeFs with negative sizes as invalid.

Fixes mumble-voip/mumble#2477