3rdparty/xinputcheck-src: add Xbox One controller GUIDs to avoid device list querying for Xbox One controllers.
Querying the device list and checking whether IG_ is in the name does
not work on Windows 10 Anniversary Update (1607).

To work around that, we add the GUIDs for all Xbox One controllers.

This should fix the problem for the most common XInput devices for now.

Presumably, Microsoft will fix this later on. But for now, this will do.

The good thing about doing it this way is that it isn't a workaround per
se -- it's an optimization that already exists in the code.

Fixes mumble-voip/mumble#2483

(cherry picked from commit 8ee2bbc51c461a42a3a6edfee6c7cee607930b17)