plugins/mumble_plugin.h: replace use of HWND with void *.
The mumble_plugin.h header is used in more places than on Windows,
but the HWND type is Windows-specific.

This has caused problems in the past, forcing all users of mumble_plugin.h
to define a fake HWND type.

We recently began passing 0 to the about() and config() plugin functions
unconditionally -- so they are practically unused.

Therefore, it makes no sense to use the HWND type anymore.
Because of that, this commit changes the HWND type to void *.
It is ABI compatible with HWND.