Log: introduce new 2048x2048 message area size restriction.
Currently, Mumble has a restrictive message filter
that will filter away messages that are larger
than the screen. A filtered message is replaced with
the text "Text object too large to display".

This commit replaces the existing size filter.
Instead of using the screen size to determine
whether a message is too large to show, Mumble
now checks whether the area of the received
message exceeds 2048x2048.

For example, this will allow messages with
sizes such as 500x8388 or 1500x2796 -- or
simply 2048x2048. As long as the total area
of the rendered message does not exceed
2048x2048 -- in which case the
"Text object too large to display" will
be displayed.

Fixes mumble-voip/mumble#2467