Add MUComboBox.
This adds a new QComboBox subclass called MUComboBox
to widgets/MUComboBox.{cpp,h}.

This QComboBox subclass explicitly uses a QListView
as the backing view for the QComboBox.

This fixes various styling issues for QComboBox
on macOS.

By default on macOS, QComboBoxes are backed by
something that tries to emulate a native macOS
menu. However, that QAbstractItemView behaves
inconsistently when styled. For example, it does
not seem possible to set the size of individual
items, because they're restricted to the height
of a normal macOS menu item.
Also, in some cases (such as the QComboBox used
for the transmission picker in MainWindow's
QToolbar),  the height of the QAbstractItemView
was also wrong when styled. This caused the combo
box to always scroll, even though it seemingly
was sized correctly.

To get consistent behavior, we use QListView as
the QComboBox's view in MUComboBox.

Also, at least for the default Mumble themes,
I've observed some weird eliding behavior for
the text of added items in the transmission picker
QComboBox in MainWindow's toolbar. Because of that,
for MUComboBox, we don't show ellipses by default.
This fixes the display of the combo box in the
MainWindow's toolbar.