MainWindow: avoid initial setupView(false) call in setupGui() on Qt 5.
The old behavior causes the MainWindow to not appear on Linux when
using certain Qt 5 versions.

Removing the call to setupView(false) fixes the problem for all
Linux users who have tried it.

I've also tested removing the code path on Windows, and everything
still works as expected.

We'll keep the code path for Qt 4 builds, though!

Fixes mumble-voip/mumble#2192