Fix #2584 by reverting introduction of log message with PR #2580
The logic does not handle all cases. Specifically, it does not handle
the normal connect case correctly (which is a "not found" case as well)

When working on a fix, more changes seemed reasonable, which resulted
in PR #2590 which is not trivial and not ready yet.

In order for the issue #2584 to not persist longer than necessary, revert
the change that introduced it. The PR #2590 will reintroduce the log
message and other changes later on.

This reverts commit 9c5d02cc0e92807348405aa5e071e7d2c5c4cd4d
(merge commit 4add9cecb9c621e02d53f5cf49719698786d3338)