compiler.pri: improve modern C++ detection.
This commit updates the modern C++ check in compiler.pri
to consider CONFIG(c++14) and CONFIG(c++1z) as "modern C++"
alongside the existing check for CONFIG(c++11).

In addition to those, we now also check if the Qt version
is 5.7 or later. Since Qt 5.7 requires at least C++11 support
to build and run, we automatically consider Qt 5.7 or later
to support modern C++ regardless of which c++ CONFIG values
are present. This is mostly to future-proof the code path
against new C++ standards that we don't check for yet, because
current Qt versions have no CONFIG flags for them.

Fixes mumble-voip/mumble#2600