compiler.pri, cplusplus.pri: improve modern C++ detection.
This commit adds a cplusplus.pri file that allows users
to pass 'c++03', 'c++11', 'c++14' or 'c++1z' in CONFIG
when building Mumble via qmake.

For example, to explicitly build Mumble in c++1z mode,
one would use:

    $ qmake -recursive -Wall main.pro CONFIG+="release c++1z"

While it is possible to explicitly specify a target C++ standard
when building Mumble, the cplusplus.pri will auto-detect an
appropriate C++ standard to target for the system it is building
on.