GlobalShortcutWin: ensure winhook NumLock and Pause keys are translated into their correct DirectInput key IDs.
When handling low-level Windows key events, we emit them to the
GlobalShortcut engine using the same key ID that we use for
the DirectInput backend.

Unfortunately, the existing code for generating a DirectInput
compatible key ID didn't produce correct results for NumLock
and Pause.

This commit fixes that by special-casing the translation of the
"extended key" flag when translating a scancode into a DirectInput
key ID.

Now, we ensure that the extended flag bit is ALWAYS set for the Pause
key (even though it isn't in the low-level key event). We also ensure
that the extended flag bit is NEVER set for the NumLock key (even though
it is in the low-level key event).

This yields correct DirectInput key IDs for NumLock and Pause keys.

Fixes mumble-voip/mumble#1353