License: new class for easier access to license data.
This new class wraps the existing "licenses.h"-header,
but makes it easier to use.

The existing "licenses.h"-header uses static string
literals, which means that if two source files include
the file, both object files will include the string data.

This License file is introduced to avoid that problem,
but also to provide a more Qt-like API, with QStrings
instead of C-style char *s.