Murmur: Add AboutDialogOptions flag to the AboutDialog class.
This flag allows users of AboutDialog to determine which
tab of the about dialog (About, License, Authors,
Third-party Licenses) should be active when the dialog
is first shown.

This is needed for the license command-line flags on Windows,
where qFatal() output is shown via Windows's ShowMessageDialog.
Unfortunately, both the AUTHORS file and the combined
third-party license text are too big to display in a regular
dialog.

To avoid this, we'll show the AboutDialog when --license,
--authors and --third-party-licenses are passed to Murmur
on Windows.

This commit only implements the AboutDialog-related changes
necessary to implement this.