Murmur: use About dialog for -license, -authors and -third-party-licenses on Windows.
On Windows, Murmur uses ShowMessageDialog for qFatal() output.

This doesn't work for -license, -authors and -third-party-licenses,
because the AUTHORS and third-party license texts are too big to
show in a regular dialog

To fix this, we instead pop up Murmur's AboutDialog when
any of the license-related command-line flags are passed
to Murmur.