3rdparty/qqbonjour-src: in BonjourServiceResolver, always tear down ResolveRecord on error.
This fixes mumble-voip/mumble#2223

Before this commit, BonjourServiceResolver would not consistently
tear down the ResolveRecord when an error occurred.

The 100% CPU usage problem from issue 2223 happens in
BonjourServiceResolver::bonjourResolveReply, where previously,
the ResolveRecord was removed from qmResolvers for both error
and success -- but was never deleted on error. To fix this issue,
we delete the ResolveRecord for bonjourResolveReply's error case.

For constency, we also introduce similar code in the
BonjourServiceResolver::bonjourSocketReadyRead method.
Now, if DNSServiceProcessResult returns an error, we tear down
the ResolveRecord and remove it from qmResolvers.