Settings: add workaround for macOS "@Variant(" bug.
Qt expects to be able to write NUL bytes in strings
serialized via QSettings. macOS Sierra broke this.

We've since backported a fix from Qt master into
our buildenv, and the bug is fixed.

This commit adds a workaround for users whose clients
have already written 'bad' config files on macOS.

If a bad config value is detected, it is skipped and
the default is used instead.