qt.pri: introduce QT_VERSION_INT for easier version comparisons in qmake files.
This introduces a new variable, QT_VERSION_INT, which allows
qmake files to more easily do Qt version checks.

Previosly, code would chain greaterThan/equals/lessThan with
logical ands for QT_MAJOR_VERSION and QT_MINOR_VERSION.

That worked OK, but was a little verbose, and is inconvenient
if you want to use a QT_VERSION check alongside other conditions
in a conditional.

The QT_VERSION_INT is a string variable is of the form:

  Qt 5.0.0:   50000
  Qt 5.6.1:   50601

it is base 10, and can be used with qmake's comparison operators
like so:

  greaterThan(QT_VERSION_INT, 50600) {
    // Do stuff for Qt > 5.6
  }

This commit also removes our own copies of QT_MAJOR_VERSION
and QT_MINOR_VERSION. These are provided by mkspecs/qconfig.pri,
and have been since (at least) Qt 4. So they're already available.
We don't need to provide them. Our method of fetching them was also
buggy due to Qt 4 compatibility - so that's fixed now.